refactor: centralize logic in notification_service
