--- Nisemono/backend/node_services/auth_service/src/core/redisClient.js
@@ -1,14 +1,0 @@
-require("dotenv").config();
-
-const redis = require("redis");
-
-const client = redis.createClient({
-  url: process.env.REDIS_URL,
-});
-
-client
-  .connect()
-  .then(() => console.log("Connected to Redis"))
-  .catch((err) => console.error("Redis connection error:", err));
-
-module.exports = client;
