--- Nisemono/backend/node_services/auth_service/src/utils/email.js
@@ -1,25 +1,0 @@
-const nodemailer = require("nodemailer");
-
-async function sendEmail(to, subject, html) {
-  const transporter = nodemailer.createTransport({
-    host: process.env.SMTP_HOST,
-    port: process.env.SMTP_PORT || 587,
-    secure: Number(process.env.SMTP_PORT) === 465,
-    auth: {
-      user: process.env.SMTP_USER,
-      pass: process.env.SMTP_PASS,
-    },
-  });
-
-  const info = await transporter.sendMail({
-    from: `"Nisemono" <${process.env.SMTP_USER}>`,
-    to,
-    subject,
-    html,
-  });
-
-  console.log("Email sent:", info.messageId);
-  return info;
-}
-
-module.exports = { sendEmail };
